engine: Move SliceTransform impls into engine_rocks::util
